Revert "some minor improves"
