--- themes/Blue/config.json
@@ -1,3 +1,0 @@
-{
-    "fileName": "Blue.css"
-}
